requirements.txt
# Runtime dependencies for the Accident Risk Prediction project.
# Keep this list lean to avoid unnecessary bloat in production environments.
pandas>=2.2.0
numpy>=1.26.0
pyyaml>=6.0.1
matplotlib>=3.8.0
pandas-stubs>=2.2.0.240218
types-requests>=2.32.0.20240907
# Geospatial/time utilities can be added when needed:
# pyproj
# shapely
# python-dateutil
